I2C protocol implemented
Change the I2C configuration of the SMBus to a normal configuration. implementation of information handling functions.
it is necessary to change the object of the sensors so that the information can be sent more efficiently, a list must be created that only has the data from the sensors, and another to indicate the channels