Feature 4 implemented and fixed
It was found that when creating the uint8_t vector for message 4, it was being created incorrectly, since it had 14 as its base size, and it had to be 16, this would generate that when trying to write the last point, lock up the whole system.
In addition, the volatile operator was added to these messages, so that problems are not found when being asked for several tasks.
In order not to arm the message in the interruption, each value in the tasks is armed.
The vector is initialized with 1 at the beginning of main to confirm that this is not the possible error.